# backend/docs/swagger.yaml
basePath: /api
definitions:
  rest.Err:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/rest.ErrDetail'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  rest.ErrDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  rest.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/rest.Err'
    type: object
...
  rest.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  rest.TagCreateRequest:
    properties:
      name:
        type: string
    type: object
  rest.TagListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.TagResponse'
        type: array
      next_page_token:
        type: string
    type: object
  rest.TagRenameRequest:
    properties:
      name:
        type: string
    type: object
  rest.TagResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  rest.TaskAssigneeListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.TaskAssigneeResponse'
        type: array
    type: object
  rest.TaskAssigneeResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      project_owner:
        type: boolean
    type: object
  rest.TaskAssignmentRequest:
    properties:
      assignee_id:
        type: string
    type: object
...
  rest.TaskResponse:
    properties:
      actual_minutes:
        type: integer
      assignee_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      priority:
        type: string
      progress:
        type: integer
      project_id:
        type: string
      revision:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  rest.TaskRevisionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.TaskRevisionResponse'
        type: array
      next_page_token:
        type: string
    type: object
  rest.TaskRevisionResponse:
    properties:
      actual_minutes:
        type: integer
      assignee_id:
        type: string
      changed_at:
        type: integer
      changed_by:
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        type: string
      due_date:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      priority:
        type: string
      project_id:
        type: string
      revision:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Task scheduling API server.
  title: task2todaytodo API
  version: "1.0"
paths:
...
  /tags:
    get:
      parameters:
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Opaque next page token
        in: query
        name: page_token
        type: string
      - description: Response field mask
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TagListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Tag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TagCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rest.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - Tags
  /tags/{id}:
    delete:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - Tags
    get:
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TagResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Get tag
      tags:
      - Tags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: New tag name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TagRenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Rename tag
      tags:
      - Tags
...
  /tasks/{id}/assignees:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskAssigneeListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List task assignees
      tags:
      - Tasks
    patch:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Assignee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TaskAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: Precondition Required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Assign task
      tags:
      - Tasks
  /tasks/{id}/revisions:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Opaque next page token
        in: query
        name: page_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskRevisionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List task revisions
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"